feat discord: Switch from Discord4J -> JDA
JDA makes it easier to implement custom requests towards the Discord
API, which is necessary for Discordian in order to invite users to a
server.

This changes from Discord4J -> JDA and also adds a new call for adding
members to an existing guild / server.

On a sidenote, the PermissionHelper now has to implement the reducing of
permissions into an API-compatible integer on its own. That's now
implemented as well as a test for it.